requirements.txt
discord.py[voice]>=2.0.0
PyQt6>=6.2.0
PyQt6-Qt6>=6.2.0
PyQt6-sip>=13.2.0
pydantic>=1.9.0
python-dotenv>=0.19.0
pynacl>=1.4.0
ffmpeg-python>=0.2.0
sounddevice>=0.4.4
soundfile>=0.10.3
numpy>=1.21.0
configparser>=5.0.0
python-rtmidi>=1.4.0
